# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "detent API",
    "description": "PLAN §2.6/§2.7: the detent-web `/api/v1` surface.",
    "license": {
      "name": "BSD-3-Clause",
      "identifier": "BSD-3-Clause"
    },
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/v1/commits/{id}/confirm": {
      "post": {
        "tags": [
          "commits"
        ],
        "summary": "`POST /api/v1/commits/{id}/confirm`.",
        "operationId": "confirm",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Commit id `Apply` returned",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The commit is now final",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommitConfirmedView"
                }
              }
            }
          },
          "409": {
            "description": "No such commit is pending",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/commits/{id}/rollback": {
      "post": {
        "tags": [
          "commits"
        ],
        "summary": "`POST /api/v1/commits/{id}/rollback`.",
        "operationId": "rollback",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Commit id `Apply` returned",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Every write since the commit was armed is undone",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RolledBackView"
                }
              }
            }
          },
          "409": {
            "description": "No such commit is pending",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CommitConfirmedView": {
        "type": "object",
        "description": "Answer to `ConfirmCommit`.",
        "required": [
          "commit_id"
        ],
        "properties": {
          "commit_id": {
            "type": "integer",
            "format": "int32",
            "description": "The commit that is now final.",
            "minimum": 0
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "The JSON body of every API failure.",
        "required": [
          "code",
          "message_id"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "A stable machine-readable class, derived from the status."
          },
          "diagnostics": {
            "type": "object",
            "description": "Validation findings, only present for a rejected [`Operation::Apply`]\n(PLAN §2.5: any `Severity::Error` diagnostic stops an apply before\nanything is written). Mirrors `detent_core::diag::Diagnostics`, which\ncannot derive `ToSchema` without giving `detent-core` a dependency on\n`utoipa`.\n\n[`Operation::Apply`]: detent_ops::Operation::Apply"
          },
          "message_id": {
            "type": "string",
            "description": "The Fluent id a front end renders."
          }
        }
      },
...
      "RolledBackView": {
        "type": "object",
        "description": "Answer to `RollbackCommit`.",
        "required": [
          "commit_id",
          "restored"
        ],
        "properties": {
          "commit_id": {
            "type": "integer",
            "format": "int32",
            "description": "The commit that was rolled back.",
            "minimum": 0
          },
          "restored": {
            "type": "integer",
            "format": "int32",
            "description": "How many targets were actually restored.",
            "minimum": 0
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_token": {
        "type": "http",
        "scheme": "bearer"
      },
      "session_cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-detent_session"
      }
    }
  },
  "security": [
    {
      "session_cookie": []
    },
    {
      "bearer_token": []
    }
  ],
...
}